[BETA] 4.3.8
Started working on an "expand browser" function for game detail windows.
Started working on a possible in-game overlay.

You can launch Windows 10 games now. Make a desktop shortcut of the game then just do the same thing you do for .exe's but link it to the shortcut file (.lnk instead of .exe)

+ You can now launch Windows 10 games.

Fixes
==============================
+ Windows 10 games and