[ALPHA] 2.0.0
Started working on a database conversion tool to convert the old .accdb format to an .mdf format. It will be connected to IGDB.com to utilize it's database on games to make filling out entries much faster. As for UGame itself...

[Main Form]
- You can now add entries.
- Your folder can now be anywhere on your computer and the database will connect.
- Added an icon column to the DGV.
- You can now press Enter to open a new tab.

[Game Tab]
- Need to fix the connection string.

... And those are the patch notes. I didn't cover everything because most of it is just bringing stuff over from UGame 1.0.

However, I do like the change I've made with URLs and Launch codes. They're a table that you can edit directly now. The table will be improved in the future, for example I'll bring the improved launch code form back. Settings will also be it's own class.